+++ spec/lib/karafka/app_spec.rb
@@ -17,7 +17,14 @@ RSpec.describe Karafka::App do
     it 'expect to run setup_components' do
       expect(Karafka::Setup::Config).to receive(:validate!).once
       expect(Karafka::Setup::Config).to receive(:setup_components).once
-      expect(Karafka::Callbacks).to receive(:after_init).once
+
+      app_class.boot!
+    end
+
+    it 'expect to publish after_init event' do
+      expect(Karafka.events).to(
+        receive(:publish).with('after_init', instance_of(Hash))
+      )
 
       app_class.boot!
     end