+++ spec/lib/karafka/consumers/includer_spec.rb
@@ -4,6 +4,7 @@ RSpec.describe Karafka::Consumers::Includer do
   subject(:includer) { described_class.new }
 
   let(:consumer_class) { Class.new(Karafka::BaseConsumer) }
+  let(:consumer) { consumer_class.new(topic) }
   let(:features_scope) { Karafka::Consumers }
   let(:backends_scope) { Karafka::Backends }
   let(:topic) do
@@ -15,16 +16,14 @@ RSpec.describe Karafka::Consumers::Includer do
     )
   end
 
-  before { consumer_class.topic = topic }
-
   describe 'inline with batch consuming' do
     let(:backend) { :inline }
     let(:batch_consuming) { true }
     let(:responder) { nil }
 
-    it { expect(consumer_class.include?(backends_scope::Inline)).to eq true }
-    it { expect(consumer_class.include?(features_scope::SingleParams)).to eq false }
-    it { expect(consumer_class.include?(features_scope::Responders)).to eq false }
+    it { expect(consumer.singleton_class.include?(backends_scope::Inline)).to eq true }
+    it { expect(consumer.singleton_class.include?(features_scope::SingleParams)).to eq false }
+    it { expect(consumer.singleton_class.include?(features_scope::Responders)).to eq false }
   end
 
   describe 'inline without batch consuming' do
@@ -32,9 +31,9 @@ RSpec.describe Karafka::Consumers::Includer do
     let(:batch_consuming) { false }
     let(:responder) { nil }
 
-    it { expect(consumer_class.include?(backends_scope::Inline)).to eq true }
-    it { expect(consumer_class.include?(features_scope::SingleParams)).to eq true }
-    it { expect(consumer_class.include?(features_scope::Responders)).to eq false }
+    it { expect(consumer.singleton_class.include?(backends_scope::Inline)).to eq true }
+    it { expect(consumer.singleton_class.include?(features_scope::SingleParams)).to eq true }
+    it { expect(consumer.singleton_class.include?(features_scope::Responders)).to eq false }
   end
 
   describe 'inline with responder' do
@@ -42,8 +41,8 @@ RSpec.describe Karafka::Consumers::Includer do
     let(:batch_consuming) { false }
     let(:responder) { Class.new }
 
-    it { expect(consumer_class.include?(backends_scope::Inline)).to eq true }
-    it { expect(consumer_class.include?(features_scope::SingleParams)).to eq true }
-    it { expect(consumer_class.include?(features_scope::Responders)).to eq true }
+    it { expect(consumer.singleton_class.include?(backends_scope::Inline)).to eq true }
+    it { expect(consumer.singleton_class.include?(features_scope::SingleParams)).to eq true }
+    it { expect(consumer.singleton_class.include?(features_scope::Responders)).to eq true }
   end
 end