+++ spec/lib/karafka/persistence/consumer_spec.rb
@@ -7,30 +7,17 @@ RSpec.describe Karafka::Persistence::Consumer do
     let(:partition) { 1 }
     let(:topic_id) { rand.to_s }
     let(:topic) do
-      instance_double(
-        Karafka::Routing::Topic,
-        persistent: persistent,
-        id: topic_id,
-        consumer: Karafka::BaseConsumer
-      )
-    end
-
-    context 'when persistence is disabled' do
-      let(:persistent) { false }
-
-      it 'expect not to cache it' do
-        r1 = persistence.fetch(topic, partition)
-        expect(r1).not_to eq persistence.fetch(topic, partition)
+      Karafka::Routing::Topic.new(
+        rand,
+        Karafka::Routing::ConsumerGroup.new(rand)
+      ).tap do |topic|
+        topic.consumer = Class.new(Karafka::BaseConsumer)
       end
     end
 
-    context 'when persistence is enabled' do
-      let(:persistent) { true }
-
-      it 'expect to cache it' do
-        r1 = persistence.fetch(topic, partition)
-        expect(r1).to eq persistence.fetch(topic, partition)
-      end
+    it 'expect to cache it' do
+      r1 = persistence.fetch(topic, partition)
+      expect(r1).to eq persistence.fetch(topic, partition)
     end
   end
 end