+++ spec/lib/karafka/setup/config_spec.rb
@@ -18,7 +18,7 @@ RSpec.describe Karafka::Setup::Config do
 
   describe '#validate!' do
     context 'when configuration has errors' do
-      let(:error_class) { ::Karafka::Errors::InvalidConfiguration }
+      let(:error_class) { ::Karafka::Errors::InvalidConfigurationError }
       let(:error_message) { { client_id: ['must be filled'] }.to_s }
 
       before do
@@ -31,7 +31,7 @@ RSpec.describe Karafka::Setup::Config do
         end
       end
 
-      it 'raise InvalidConfiguration exception' do
+      it 'raise InvalidConfigurationError exception' do
         expect { config_class.send(:validate!) }.to raise_error do |error|
           expect(error).to be_a(error_class)
           expect(error.message).to eq(error_message)
@@ -50,7 +50,7 @@ RSpec.describe Karafka::Setup::Config do
     end
 
     context 'when configuration is valid' do
-      it 'not raise InvalidConfiguration exception' do
+      it 'not raise InvalidConfigurationError exception' do
         expect { config_class.send(:validate!) }
           .not_to raise_error
       end